PR review (renaming files) and fixed legend tests
