+++ v3/cypress/fixtures/attribute-types.codap
@@ -1,5 +1,5 @@
 {
-	"name": "attribute-types-test-document",
+	"name": "attribute-types",
 	"guid": 1,
 	"id": 1,
 	"components": [{
