+++ v3/cypress/fixtures/table-groups.codap
@@ -1,5 +1,5 @@
 {
-    "name": "3TableGroups",
+    "name": "table-groups",
     "guid": 1,
     "components": [
         {
