+++ v3/cypress/support/elements/legend-elements.ts
